Parked vehicles + split NavigationGrid into nav + CellTopology
Vehicles: VehicleKind enum (UTILITY_TRUCK / FLATBED_TRUCK / TANKER_TRUCK
/ BOX_VAN from trucks.png; CARGO_TRUCK / ARMORED_APC from trucks_2.png),
each bundling sheet + frame index + 3x2 footprint. BattleSetup stamps 3-6
on random street/courtyard cells before sim construction, marks footprint
cells non-walkable, and recomputes the cover ring so adjacent units pick
up the cover bonus. BattleScreen lazy-loads both sheets, fills each
footprint with road color, and draws the sliced truck sprite scaled to
fit the footprint while preserving aspect.

Cell tagging refactor: NavigationGrid's cellFlags widens from byte[] to
long[] and exposes hasTag(x,y,CellTag) / setTag for generic access.
Pathfinder hot-path still does `& 1L` against WALKABLE at ordinal 0.

Topology split: the rendering / categorization tags (FLOOR, RUBBLE,
STREET, CROSSWALK, CROSSWALK_HORIZ, COURTYARD, WALL, VEHICLE) move out
of NavigationGrid into a new CellTopology class at battle/map/. The
nav grid keeps WALKABLE + DOORWAY (zone graph) plus edges / cover /
wallHp; topology owns everything the renderer or placement filters
read. UrbanMapGenerator threads topology through placeBuilding /
dissolveMegaInteriors / paintCrosswalks and ends with
topology.tagDefaultWalls(grid). BattleSimulation.damageCell flips
topology.setWall(false) / setRubble(true) / setFloor(true) after the
nav grid mutates. The renderer's wall pass iterates topology.isWall
directly so vehicle cells (non-walkable, not WALL) skip wall art with
no per-call guard.

Also includes the user's parallel BattleCamera extraction at
ops/battleview/, picked up by BattleScreen + FlybyOverlay since they
were already mixed-touched in this branch.

+++ src/main/java/com/dillon/starsectormarines/battle/nav/GridPathfinder.java
@@ -234,7 +234,7 @@ public final class GridPathfinder {
         heapPos[startIdx] = 0;
         int heapSize = 1;
 
-        byte[] cellFlags = grid.getCellFlagsArray();
+        long[] cellFlags = grid.getCellFlagsArray();
         byte[] edgePass  = grid.getEdgePassabilityArray();
 
         while (heapSize > 0) {
@@ -264,7 +264,7 @@ public final class GridPathfinder {
 
                 int nIdx = ny * w + nx;
 
-                if ((cellFlags[nIdx] & 1) == 0) continue;
+                if ((cellFlags[nIdx] & 1L) == 0L) continue;
                 // Dual-side edge check: source cell's edge AND destination's reciprocal.
                 if ((edgePass[currentIdx] & DIR_EDGE_MASK[dirI]) == 0) continue;
                 if ((edgePass[nIdx]       & DIR_OPP_EDGE_MASK[dirI]) == 0) continue;
@@ -282,8 +282,8 @@ public final class GridPathfinder {
                     int a2y = cy + DIR_ADJ2_DY[dirI];
                     if (a1x < 0 || a1x >= w || a1y < 0 || a1y >= h) continue;
                     if (a2x < 0 || a2x >= w || a2y < 0 || a2y >= h) continue;
-                    if ((cellFlags[a1y * w + a1x] & 1) == 0) continue;
-                    if ((cellFlags[a2y * w + a2x] & 1) == 0) continue;
+                    if ((cellFlags[a1y * w + a1x] & 1L) == 0L) continue;
+                    if ((cellFlags[a2y * w + a2x] & 1L) == 0L) continue;
                 }
 
                 if (heapPos[nIdx] == CLOSED) continue;
